+++ magicarp/plugins/__init__.py
@@ -1,0 +1,5 @@
+"""Plugins are optional elements of magicarp, they tend to be on/off depending
+on values set in settings files.
+"""
+
+from . import rollbar  # NOQA