+++ magicarp/tools/__init__.py
@@ -8,7 +8,7 @@ be separated to own namespace.
 """
 
 # first import all the modules that do not have any other dependency
-from . import datetime_helpers, api_request, logging, auth_model  # NOQA
+from . import datetime_helpers, api_request, auth_model  # NOQA
 
 # then all others that are actually dependening on other modules
 from . import helpers  # NOQA